Error handled in Brand

+++ frontend/rentACar/src/redux/utilities/interceptors/axiosInterceptors.js
@@ -1,9 +1,5 @@
 import axios from "axios";
-import {
-  toastError,
-  toastSuccess,
-  toastWarning,
-} from "../../../service/ToastifyService";
+
 
 const axiosInstance = axios.create({
   baseURL: "http://localhost:8080/",
@@ -19,7 +15,7 @@ axiosInstance.interceptors.request.use(
 
       return config;
     } catch (error) {
-      toastError("Request interceptor error:", error);
+      //console.error("Request interceptor error:", error);
       throw error; 
     }
   },
@@ -34,49 +30,53 @@ axiosInstance.interceptors.response.use(
   async (error) => {
     const originalRequest = error.config;
 
-    if (
-      (error.response.status === 403 || error.response.status === 401) &&
-      !originalRequest._retry
-    ) {
+    if (error.response.status === 403 || error.response.status === 401) {
+      //Sadece 403 hatası alındığında çıkış yap
+      localStorage.clear();
+      //window.location.href = "/login";
+      return Promise.reject(error);
+    }
+
+    if (!originalRequest._retry) {
       originalRequest._retry = true;
 
       try {
         const refreshToken = localStorage.getItem("refresh_token");
+        if (refreshToken) {
+          const headers = {
+            Authorization: `Bearer ${refreshToken}`,
+          };
+          const refreshResponse = await fetch(
+            "http://localhost:8080/api/v1/auth/refresh-token",
+            {
+              method: "POST",
+              headers: headers,
+            }
+          );
 
-        const headers = {
-          Authorization: `Bearer ${refreshToken}`,
-        };
-        const refreshResponse = await fetch(
-          "http://localhost:8080/api/v1/auth/refresh-token",
-          {
-            method: "POST",
-            headers: headers,
+          const refreshData = await refreshResponse.json(); // Yanıtı JSON formatına dönüştür
+
+          const newAccessToken = refreshData.access_token;
+          localStorage.removeItem("access_token");
+          localStorage.setItem("access_token", newAccessToken);
+
+          originalRequest.headers.Authorization = `Bearer ${newAccessToken}`;
+          if (originalRequest.headers.Authorization) {
+            //toastSuccess("Kullanıcının Access Tokenı Değişti");
           }
-        );
-
-        const refreshData = await refreshResponse.json(); // Yanıtı JSON formatına dönüştür
-
-        const newAccessToken = refreshData.access_token;
-        localStorage.removeItem("access_token");
-        localStorage.setItem("access_token", newAccessToken);
-
-        originalRequest.headers.Authorization = `Bearer ${newAccessToken}`;
-        if (originalRequest.headers.Authorization) {
-          toastSuccess("Kullanıcının Access Tokenı Değişti");
+          return axiosInstance(originalRequest);
+        } else {
+          throw new Error("Refresh token yok");
         }
-        return axiosInstance(originalRequest);
       } catch (refreshError) {
         // Refresh token hatası
-
-        toastError("Response interceptor Refresh token error:", refreshError);
+        //console.error("Response interceptor Refresh token error:", refreshError);
         localStorage.clear();
         window.location.href = "/login";
         return Promise.reject(refreshError);
       }
     }
-    toastWarning("Bir hata oluştu, lütfen tekrar deneyin.");
-    localStorage.clear();
-    window.location.href = "/login";
+
     return Promise.reject(error);
   }
 );
